[refactor]  move auxiliary functions to utils.py in graphs/generate/
